chore(core): align changes after router builder split and lint enforcement\n\n- Update imports/usages across tests and helpers\n- Apply one-class-per-file enforcement (no functional changes)\n- Minor formatting and consistency tweaks
