+++ packages/core/src/common/event/function-publisher.ts
@@ -24,17 +24,3 @@ export class FunctionPublisher implements EventPublisher {
     }
   }
 }
-
-export class CompositePublisher implements EventPublisher {
-  constructor(private readonly targets: readonly EventPublisher[]) {}
-
-  publish(channel: string, payload: unknown): void {
-    for (const p of this.targets) {
-      try {
-        p.publish(channel, payload);
-      } catch {
-        // isolate failures between targets
-      }
-    }
-  }
-}
