+++ packages/core/src/orchestrator/router/__tests__/perf-smoke.test.ts
@@ -2,7 +2,7 @@ import { CompositeAgentRouter } from '../composite-router';
 import { BM25TextStrategy } from '../strategies/bm25-text-strategy';
 import { KeywordBoostStrategy } from '../strategies/keyword-boost-strategy';
 import { MentionStrategy } from '../strategies/mention-strategy';
-import { EnglishSimpleTokenizer } from '../../../knowledge/tokenizer';
+import { EnglishSimpleTokenizer } from 'src/knowledge/english-simple-tokenizer';
 import type { Agent, AgentChatResult } from '../../../agent/agent';
 import type { AgentSession } from '../../../agent/agent-session';
 import type { ReadonlyAgentMetadata } from '../../../agent/agent-metadata';
